requirements.txt
streamlit==1.39.0
anthropic
python-dotenv
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-qdrant
qdrant-client
pypdf
google-genai